+++ internal/dial/cluster.go
@@ -2,7 +2,8 @@ package dial
 
 import (
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/driver/cluster"
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/driver/cluster/balancer"
+	ibalancer "github.com/ydb-platform/ydb-go-sdk/v3/internal/driver/cluster/balancer/iface"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/driver/cluster/balancer/single"
 	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
 )
 
@@ -10,11 +11,11 @@ func (d *dialer) newCluster(trace trace.Driver) cluster.Cluster {
 	return cluster.New(
 		trace,
 		d.dial,
-		func() balancer.Balancer {
+		func() ibalancer.Balancer {
 			if d.config.DiscoveryInterval() == 0 {
-				return balancer.Single()
+				return single.Balancer()
 			}
-			return balancer.New(d.config.BalancingConfig())
+			return d.config.Balancer()
 		}(),
 	)
 }